requirements.txt
fastapi==0.99.1
google-api-python-client==2.86.0
redis==4.5.5
numpy>=1.26.0
pandas==2.2.3
reportlab==4.0.0
openpyxl==3.1.2
uvicorn==0.22.0
pypdf2==3.0.1
matplotlib==3.10.3
python-dotenv==1.0.0
pydantic==1.10.15
httpx==0.24.1
gunicorn==20.1.0
nodeenv==1.8.0
